Theory nodes that re-appear in subsequent runs are displayed correctly
I.e. Imagine a Theory powered by a property returning random data. In run 1, a
value of x is passed in, the dynamic element is created and displayed in the UI.
In run 2, x isn't used, so at the end of run 2, the element is marked invalid.
At the start of run 3, the invalid elements are removed. If value x happens again
in run 3 or later, the invalid element would not get added back into the UI. It
is now.