Stop creating unit test elements with a method name of "???"
This happens when the method doesn't yet have a declaration name, so the PSI
system uses the special name of "???". I was creating a method element, adding
it to the parent class element and then (due to the lack of a declaration) it
would not get submitted to the unit test system. But when it came time to run
the tests for the class, it would try and run a method called "???" and xunit
would rightly fail. No longer creates an element when given a method called
"???" (same behaviour as the nunit test runner)

Work item: 4142