Nested classes now display the parent classes as part of the name.
Not sure this is the nicest way of doing it - the text is "Parent+Nested", which
is a little technical (you only usually see the "+" when you're dealing with
reflection.

Perhaps in future there should be a node for the parent class? I'll wait until
resharper implements support for nunit and see how they do it.
See http://youtrack.jetbrains.com/issue/RSRP-199079

I've also added an indicator for any generic types (e.g. "Class`2") to make it a
bit more visible, especially as xunit doesn't support them (or rather, it treats
them as valid test classes, but fails to instantiate them).

work item: 4143