Refactored task runners to remove obsolete warnings
ReSharper 7.1 has made the non-recursive members of RemoteTaskRunner obsolete.
This code change cleans up that code, adds a huge explanation of why we're not
using AssemblyLoadTask and cleans up the code we need because we're not using it

Also, dotCover 2.0 code doesn't build at all. Not sure how long that's been like
that