Can now handle multiple Theories with the same parameter descriptions
For primitive types, the answer is really "don't use the same data twice". E.g.
TestMethod1(12) and TestMethod1(12) will return the same result. The fun part is
with complex types. Xunit displays the result of ToString in the parameter list,
so if that's not overridden, it's very easy to get repeated Theories. This fix
appends a count to the theory row if it sees that it's been used before, e.g.
"TestMethod1(v: System.String[])" and "TestMethod1(v: System.String[]) [2]"

Work item: 4155