Explicitly specify the config file so we don't fallback to the exe's config
If we pass null for the config file when creating a new AppDomain, we use the
AppDomain's base assembly's config file if it exists, or fallback to the current
AppDomain's config file. This causes problems running the provider's own tests,
as the test runner exe's config file tries to redirect the resharper assemblies
to the latest version, but the new AppDomain doesn't know where to get the
redirected assemblies from. Boom!