Support for traits on methods
If the trait name is category (case insensitive), the value is the category,
otherwise, the category is created in the form "name[value]". Traits defined on
classes are inherited by methods.

No code completion supported yet, and traits defined on classes not displayed in
the drop down for that class.

Work item: 4128