Exceptions always use fully qualified type names
The xunit1 API would pass an Exception instance when throwing outside of a test
(e.g. invalid parameters), and the runner would use a ReSharper API to format
the exception, which used the short name of the Exception. The xunit2 API no
longer passes an Exception instance at all, so the ReSharper API can no longer
be used. It's an arbitrary split, too - why should the name be truncated here
and not everywhere? So, all exceptions are now fully qualified