Support running a single test
When the runner performs discovery, it gets all methods and theories for
a class, but it doesn't know if a method or theory is dynamic, because
the provider only gives it the methods/theories it wants to run. So, it
errs on the side of caution and runs any other method/theory as if it
were dynamically discovered. This change means the provider tells the
runner all known methods/theories for the class. The runner will now
only run methods/theories asked for, and any unknown methods - these
actually are dynamically discovered. Doesn't cover the edge case of
non-pre-enumerated theories (e.g. xunit1 theories - we will always run
all theories, we can't filter these out). Fixes #9.